consolidated file/folder structure of individual projects into staging for tweaking
