+++ Demo/GiphyAPI/scripts/global.js
@@ -1,0 +1,1 @@
+const GIPHY_KEY = 'qcJVUTcz3nQoeWZidSzzlzPB8VBZpmrw';
