+++ Demo/MovieDB/index.html
@@ -1,0 +1,36 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <link href="https://fonts.googleapis.com/css?family=Montserrat" rel="stylesheet">
+    <!-- Movie -->
+    <link rel="stylesheet" href="/Demo/MovieDB/stylesheets/MovieDB.css">
+    <title>Movie.IO</title>
+</head>
+<body>
+
+    <div class="movie-content" data-details>
+            <div class ="smallHeader">
+                <h1>MOVIE.IO</h1>
+                <form class="form2">
+                        <input type="text">
+                        <button type="button" onclick="movieSearch();">GO!</button>
+                    </form>
+
+
+
+
+
+
+            </div>
+
+
+</div>
+
+
+<!-- MovieDB.js -->
+    <script src="/Demo/MovieDB/scripts/MovieDB.js"></script>
+</body>
+</html>
