--- giphy-app/app.css
@@ -1,86 +1,0 @@
-body{
-    padding: 50px;
-}
-
-#thumbs{
-    margin-left: -6px;
-    margin-right: -6px;
-}
-
-.img-wrapper{
-    display: table;
-    height: 140px;
-    width: 200px;
-    table-layout: fixed;
-    float: left;
-    padding-left: 6px;
-    padding-right: 6px;
-    margin-bottom: 12px;
-}
-.outer{
-    height: 100%;
-}
-.inner{
-    padding: 5px;
-    border-radius: 6px;
-    border: 1px solid blueviolet;
-    background-color: pink;
-    display: table;
-    width: 100%;
-    vertical-align: middle;
-}
-.inner img{
-    display: block;
-    margin: 0 auto;
-    max-width: 100%;
-    max-height: 128px;
-}
-
-.loader-wrapper{
-    display: none;
-    position: fixed;
-    left: 0;
-    right: 0;
-    top: 0;
-    bottom: 0;
-    z-index: 20000;
-    background:yellow;
-}
-.loader-wrapper.shown{
-    display: block;
-}
-
-.lds-circle {
-  display: inline-block;
-  transform: translateZ(1px);
-    position: absolute;
-    left: 50%;
-    top: 50%;
-    margin-left: -25px;
-    margin-top: -25px;
-}
-.lds-circle > div {
-  display: inline-block;
-
-  width: 51px;
-  height: 51px;
-  margin: 6px;
-  border-radius: 50%;
-  background: hotpink;
-  animation: lds-circle 2.4s cubic-bezier(0, 0.2, 0.8, 1) infinite;
-}
-@keyframes lds-circle {
-  0%, 100% {
-    animation-timing-function: cubic-bezier(0.5, 0, 1, 0.5);
-  }
-  0% {
-    transform: rotateY(0deg);
-  }
-  50% {
-    transform: rotateY(1800deg);
-    animation-timing-function: cubic-bezier(0, 0.5, 0.5, 1);
-  }
-  100% {
-    transform: rotateY(3600deg);
-  }
-}
