--- giphy-app/index.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Giphy </title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
-          integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-    <link rel="stylesheet" href="app.css">
-</head>
-<body>
-
-<div class="container">
-    <form id="searchForm">
-        <div class="input-group mb-3">
-            <input id="searchInput" type="text" class="form-control" placeholder="Search for a gifs"
-                   aria-label="Recipient's username"
-                   aria-describedby="button-addon2">
-            <div class="input-group-append">
-                <button class="btn btn-outline-secondary" id="searchButton">Search</button>
-            </div>
-        </div>
-    </form>
-
-    <div id="thumbs">
-
-    </div>
-   
-        
-</div>
-<div class="loader-wrapper">
-    <div class="lds-circle"><div></div></div>
-</div>
-<script
-        src="https://code.jquery.com/jquery-3.3.1.js"
-        integrity="sha256-2Kok7MbOyxpgUVvAk/HJ2jigOSYS2auK4Pfzbm7uH60="
-        crossorigin="anonym ous"></script>
-<script src="scripts/global.js"></script>
-<script src="scripts/scripts.js"></script>
-</body>
-</html>
