--- giphy-app/scripts/global.js
@@ -1,1 +1,0 @@
-const GIPHY_KEY = 'qcJVUTcz3nQoeWZidSzzlzPB8VBZpmrw';
