--- index.html
@@ -1,82 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-<!-- Main Index.Html CSS -->
-    <link rel="stylesheet" href="stylesheets/index.css">
-
-<!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
-    integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-
-<!-- GiphyApp CSS -->
-    <link rel="stylesheet" href="/giphy-app/app.css">
-
-<!-- Google Fonts API -->
-    <link href="https://fonts.googleapis.com/css?family=Montserrat" rel="stylesheet">
-
-    <title>Mio is [M]ovie.[io]</title>
-   </head>
-   <body>
-       <div class ="header">
-            <h1>Movie.io</h1>
-           <h2>MioDB | MioGifs | MioKb | Mio Advisors</h2>
-           
-           <form class="form">
-               <input type="text">
-               <button type="button" onclick="movieSearch();">Go Mio!</button>
-           </form>
-       </div>
-
-       <a href="tel:+14044456860">Call a Mio Movie Advisor Now! </a>
-
-                   <div class="fullscreen-bg">
-                           <video loop muted autoplay poster="img/videoframe.jpg" class="fullscreen-bg__video">
-                               <source src="assets/Oscars-90-years-montage-edited.mp4" type="video/mp4">
-                           </video>
-                   </div>
-
-           <div class="text-content" data-details>
-               
-           </div>
-           <div></div>
-
-           <div class="container">
-            <form id="searchForm">
-                <div class="input-group mb-3">
-                    <input id="searchInput" type="text" class="form-control" placeholder="Search for a gifs"
-                           aria-label="Recipient's username"
-                           aria-describedby="button-addon2">
-                    <div class="input-group-append">
-                        <button class="btn btn-outline-secondary" id="searchButton">Go Mio</button>
-                    </div>
-                </div>
-            </form>
-        
-            <div id="thumbs">
-        
-            </div>
-           
-        <!--  -->
-        </div>
-        <div class="loader-wrapper">
-            <div class="lds-circle"><div></div></div>
-        </div>
-        <script
-                src="https://code.jquery.com/jquery-3.3.1.js"
-                integrity="sha256-2Kok7MbOyxpgUVvAk/HJ2jigOSYS2auK4Pfzbm7uH60="
-                crossorigin="anonymous"></script>
-    
-    <!-- giphy-app script tags -->
-    <script src="giphy-app/scripts/global.js"></script>
-    <script src="giphy-app/scripts/scripts.js"></script>
-
-    <!-- Main index.html tags -->
-
-   <script src="scripts/index.js"></script>
-</body>
-</html>
