--- results.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link href="https://fonts.googleapis.com/css?family=Montserrat" rel="stylesheet">
-    <link rel="stylesheet" href="stylesheets/results.css">
-    <title>Movie.IO</title>
-</head>
-<body>
-
-    <div class="movie-content" data-details>
-            <div class ="smallHeader">
-                <h1>MOVIE.IO</h1>
-                <form class="form2">
-                        <input type="text">
-                        <button type="button" onclick="movieSearch();">GO!</button>
-                    </form>
-
-
-
-
-
-                    
-            </div>
-    
-    
-</div>
-
-
-
-    <script src="scripts/results.js"></script>
-</body>
-</html>
