--- stylesheets/index.css
@@ -1,107 +1,0 @@
-/* html {
-    box-sizing: border-box;
-  }
-  *, *:before, *:after {
-    box-sizing: inherit;
-  }
-
-body {
-    background: rgb(242, 242, 242);
-}
-
-.fullscreen-bg {
-    z-index: -1;
-    top: 0;
-    bottom: 0;
-    width: 100%;
-    height: 100%;
-    overflow: hidden;
-}
-
-.fullscreen-bg__video {
-    min-width: 100%;
-    min-height: 100%;
-    width: auto;
-    height: auto;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%,-50%);
-}
-
-.header {
-    text-align: center;
-    color: white;
-    font-family: 'Montserrat', sans-serif;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    margin-top: 200px; 
-    z-index: 1000;
-    position: sticky;
-}
-
-.header h1 {
-    font-size: 54px;
-    font-weight: 20px;
-}
-
-.form {
-    display: flex;
-    justify-content: center;
-}
-
-.form2 {
-    display: flex;
-    justify-content: center;
-}
-
-input {
-    height: 20px;
-    font-family: 'Montserrat';
-    background-color: rgb(256, 256, 256, 50%);
-    padding: 20px;
-    border-radius: 50px;
-    font-size: 24px;
-    border-style: none;
-    color: white;
-    text-align: center;
-}
-
-input:focus {
-    outline:none;
-}
-
-button {
-    width: 40px;
-    height: 40px;
-    font-family: 'Montserrat';
-    background-color: rgb(255, 0, 0);
-    border-radius: 100%;
-    font-size: 14px;
-    border-style: none;
-    color: white;
-    text-align: center;
-}
-
-.text-content {
-    width: 100%;
-    height: 100%;
-}
-
-.smallHeader {
-    color: black;
-    font-family: 'Montserrat', sans-serif;
-    display: flex; 
-    justify-content: flex-start;
-    align-items: center;
-    margin: 10px;
-    font-weight: 30px;
-}
-
-.smallHeader input {
-    width: 125px;
-    margin-left: 30px;
-    background-color: red;
-}
- */
